Integration Test Suite (#883)
Creates a new testing suite called IntegrationTest that generalizes instance setups and deployments. The testing suite is an abstract contract that is implemented in instance test. The implementor overrides two virtual functions and the testing suite properly configures the forked environment. The suite then provides testing coverage for the deployment and initializing steps of the market creation process.